pyproject.toml
[project]
name = "mimifuku"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-api-python-client>=2.169.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
]

[dependency-groups]
dev = [
    "mypy>=1.15.0",
    "ruff>=0.11.10",
]

[[tool.mypy.overrides]]
module = [
    'google.auth.transport.requests',
    'google.oauth2.credentials',
    'google_auth_oauthlib.flow',
    'googleapiclient.errors',
    'googleapiclient.discovery',
    'googleapiclient.http',
]
ignore_missing_imports = true